Controller updated, requests and handler consolidated marketReuests folder name change
