+++ Application/Requests/Market/CheckMarketCodeExistsQuery.cs
