+++ Application/Requests/Market/CheckMarketNameExistsQuery.cs
