+++ Application/Requests/Market/DeleteMarketByIdCommand.cs
