+++ Application/Requests/Market/GetMarketByIdQueryHandler.cs
