+++ Application/Requests/Market/GetMarketDetailsByIdQuery.cs.cs
