+++ Application/Requests/Market/GetMarketDetailsByIdQueryHandler.cs.cs
