+++ Application/Requests/Market/SearchMarketQuery.cs
