+++ Application/Requests/Market/SearchMarketQueryHandler.cs
